Dashboard activity is added. Download Entries activity is added. Some beautification is done in dashboard and other activities. Version 2.0.
Download Entries activity ask for start date and end date and it will download the entries from start date upto end date in excel file.
If start date or end date is not in correct format it prompts associated error to the user.
If start date or end date entry is not found in database than also it prompts associated error to the user i.e. No Entry found for this date.
If login ID and/or Password is not correct than also it prompts error.
Previous functionality is not removed till now means if we want to clear databse we should go to Main2Activity and from App Bar we can clear
all the entries from database also we can view the entries from app bar.